+++ examples/replication.py
@@ -1,3 +1,5 @@
+import asyncio
+
 import pytest
 
 from asyncmy import connect
@@ -6,21 +8,23 @@ from asyncmy.replication.row_events import WriteRowsEvent
 from conftest import connection_kwargs
 
 
-@pytest.mark.skip(reason="need test in local")
-@pytest.mark.asyncio
-async def test_binlogstream(connection):
+async def main():
     conn = await connect(**connection_kwargs)
+    ctl_conn = await connect(**connection_kwargs)
 
     stream = BinLogStream(
-        connection,
         conn,
+        ctl_conn,
         1,
-        master_log_file="binlog.000172",
+        master_log_file="binlog.000019",
+        master_log_position=155276,
         resume_stream=True,
         blocking=True,
-        master_log_position=2235312,
     )
-    await stream.connect()
     async for event in stream:
         if isinstance(event, WriteRowsEvent):
             print(event.schema, event.table, event.rows)
+
+
+if __name__ == "__main__":
+    asyncio.run(main())